requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
astronomer-cosmos[dbt.snowflake]
anyio==3.7.1
apache-airflow
apache-airflow-providers-common-sql==1.6.0
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-http==4.5.0
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.2
apache-airflow-providers-dbt-cloud==2.2.0
apispec==5.2.2
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
boto3==1.27.0
botocore==1.30.0
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.1
Deprecated==1.2.14
dill==0.3.6
dnspython==2.4.0
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.2
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
flask-limiter
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.4.0
google-re2==1.0
graphviz==0.20.1
gunicorn==21.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
ordered-set==4.1.0
packaging==23.1
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.2.0
prison==0.2.1
psutil==5.9.5
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.8.0
pytest==7.4.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich>=12.4.4,<14
rich-argparse==1.2.0
rpds-py==0.9.2
s3transfer==0.6.1
setproctitle==1.3.2
sniffio==1.3.0
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.6.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.16
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.2
typer[all]>=0.6.1,<0.7.0

